frameworks/base:Enable tunnel mode encoding and Voicememo recording support (AMRNB, QCELP and EVRC codecs).
Enable tunnel mode encoding for AMRNB, QCELP, EVRC codecs. This feature will be used for tunnel mode audio recording in camcorder.
Voicememo recording support enabled for AMRNB, QCELP and EVRC codecs. Voice call (Tx path, Rx path and Tx+Rx path) can be recorded with these codecs.

+++ include/media/AudioRecord.h
@@ -336,7 +336,7 @@ private:
     uint32_t                mFrameCount;
 
     audio_track_cblk_t*     mCblk;
-    uint8_t                 mFormat;
+    uint32_t                mFormat;
     uint8_t                 mChannelCount;
     uint8_t                 mInputSource;
     uint8_t                 mReserved;
